Split gui and msgbox unit tests
